feat: rename model utils module. resolve the module dependencies.
