+++ src/facefl/client/base_client.py
@@ -23,9 +23,9 @@ from flwr.common.logger import log
 from torch.utils.data import DataLoader
 
 from facefl.dataset import configure_dataset, load_federated_dataset, split_validation
+from facefl.model import load_model
 from facefl.model.base_model import Net
 from facefl.model.driver import test, train
-from facefl.utils.utils_model import load_model
 
 warnings.filterwarnings("ignore")
