+++ src/facefl/face_verification/inference_visualizer.py
@@ -5,11 +5,10 @@ import matplotlib.pyplot as plt
 import numpy as np
 import torch
 from torch.utils.data import DataLoader
-from utils.utils_model import load_arcface_model
 from utils.utils_plot import fig_setup
 
 from facefl.dataset import load_centralized_dataset, load_federated_dataset
-from facefl.model import Net
+from facefl.model import Net, load_arcface_model
 
 
 def imshow(img):
