+++ src/facefl/local/server.py
@@ -12,9 +12,9 @@ from flwr.server.strategy import FedAvg
 from torch.utils.data import DataLoader
 
 from facefl.dataset import configure_dataset, load_centralized_dataset
+from facefl.model import load_model
 from facefl.model.base_model import Net
 from facefl.model.driver import test
-from facefl.utils.utils_model import load_model
 
 warnings.filterwarnings("ignore")
