+++ src/facefl/model/__init__.py
@@ -1,5 +1,6 @@
 from .base_model import Net
 from .driver import test, train
+from .model import load_arcface_model, load_model
 from .resnet import ResNet, ResNetLR
 from .tinycnn import tinyCNN
